pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Nikita Bormintsev <nikita.bormintsev@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.3"
uvicorn = {extras = ["standart"], version = "^0.29.0"}
asyncpg = "^0.29.0"
bcrypt = "^4.1.3"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[tool.poetry.scripts]
app = "src.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
